requirements.txt
fastapi==0.111.1
httpx
pylint
pytest-asyncio
